backend/requirements.txt
Django==4.2.23
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.9
django-redis==5.4.0
